extract ClientErrorHandler from extension.ts
