+++ src/syntaxLanguageClient.ts
@@ -4,7 +4,8 @@ import * as net from "net";
 import { DidChangeConfigurationNotification, LanguageClientOptions } from "vscode-languageclient";
 import { LanguageClient, ServerOptions, StreamInfo } from "vscode-languageclient/node";
 import { apiManager } from "./apiManager";
-import { ClientErrorHandler, getJavaConfig } from "./extension";
+import { ClientErrorHandler } from "./clientErrorHandler";
+import { getJavaConfig } from "./extension";
 import { ClientStatus, ExtensionAPI } from "./extension.api";
 import { logger } from "./log";
 import { OutputInfoCollector } from "./outputInfoCollector";
